# app_backend/swagger.yaml
openapi: 3.0.3
info:
  title: API de Gestión de Inventario, Préstamos y Penalizaciones
  description: >
    Esta API maneja la autenticación de usuarios, administración de perfiles, 
    control centralizado de stock (materiales, herramientas, placas, libros), 
    flujo de préstamos/reservas mediante validación por QR, gestión de penalizaciones 
    y generación de reportes/estadísticas.
  version: 2.0.0
servers:
  - url: /
    description: Servidor Base Local
paths:
...
  /api/articulos:
    get:
      tags:
        - Inventario y Materiales
      summary: Lista el material disponible
      description: Acceso público (sin autenticación requerida). Incluye soporte para filtros por tipo, estado y disponibilidad.
      parameters:
        - name: category
          in: query
          required: false
          description: Filtrar por tipo o categoría.
          schema:
            type: string
        - name: condition
          in: query
          required: false
          description: Filtrar por condición física.
          schema:
            type: string
        - name: available
          in: query
          required: false
          description: Filtrar solo elementos disponibles.
          schema:
            type: boolean
      responses:
        '200':
          description: Catálogo obtenido correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemDetail'

    post:
      tags:
        - Inventario y Materiales
      summary: Ingreso/creación de un nuevo ítem al inventario
      description: (Solo Admin).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Ítem agregado con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetail'
        '400':
          description: Datos incorrectos.
        '403':
          description: Prohibido. Permisos insuficientes.
  /api/articulos/{articulo_id}:
    parameters:
      - name: articulo_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Inventario y Materiales
      summary: Consulta el detalle de un ítem por ID
      description: Endpoint GET implementado en la variante de compatibilidad (tabla items).
      responses:
        '200':
          description: Detalle del ítem.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemLegacyDetail'
        '404':
          description: Ítem no encontrado.

    put:
      tags:
        - Inventario y Materiales
      summary: Actualiza la información general de un ítem
      description: (Solo Admin).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
      responses:
        '200':
          description: Ítem actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetail'
        '403':
          description: Prohibido.
        '404':
          description: Ítem no encontrado.

    delete:
      tags:
        - Inventario y Materiales
      summary: Baja definitiva o lógica del ítem del inventario
      description: (Solo Admin).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Ítem eliminado.
        '403':
          description: Prohibido.
        '404':
          description: Ítem no encontrado.
  /api/articulos/{articulo_id}/condition:
    patch:
      tags:
        - Inventario y Materiales
      summary: Actualiza específicamente el estado o condición física del ítem
      parameters:
        - name: articulo_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - estado
              properties:
                estado:
                  type: string
                  enum: [disponible, dañado, reparacion, "dado de baja"]
      responses:
        '200':
          description: Condición actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetail'
        '404':
          description: Ítem no encontrado.

  # ==========================================
  # 4. Préstamos y Solicitudes (/api/loans)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ItemDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre_art:
          type: string
        tipo:
          type: string
        seccion:
          type: string
        prestacion_maxima:
          type: integer
        stock:
          type: integer
        necesita_reparacion:
          type: boolean
    ItemLegacyDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          type: string
    ItemCreate:
      type: object
      required:
        - nombre_art
        - tipo
        - seccion
        - prestacion_maxima
      properties:
        nombre_art:
          type: string
        tipo:
          type: string
        seccion:
          type: string
        prestacion_maxima:
          type: integer
        stock:
          type: integer
        necesita_reparacion:
          type: boolean
    ItemUpdate:
      type: object
      properties:
        nombre_art:
          type: string
        tipo:
          type: string
        seccion:
          type: string
        prestacion_maxima:
          type: integer
        stock:
          type: integer
        necesita_reparacion:
          type: boolean
...